An implementation of a 4-bit Nanoprocessor
This project was done as a part of the Computer Architecture module I
followed at University of Moratuwa. The project was designed in
schematics using Xilinx ISE Design Suite. This can be implemented in an
BASYS 2 FPGA board. This nanoprocessor is capable of performing four
instructions. A program to calculate the total of integers up to 3 has
been hard coded to a ROM as a test. Each instruction is fed into the
nanoprocessor at each clock pulse. A test bench file for simulation is
also included